Refactored and added Analytics Agent startup

+++ build.sh
@@ -9,14 +9,16 @@
 #! /bin/bash
 
 cleanUp() {
-  (cd ADCapital-Tomcat && rm -f AppServerAgent.zip AnalyticsAgent.zip)
-  (cd ADCapital-Tomcat && rm -rf AD-Capital)
-  (cd ADCapital-ApplicationProcessor && rm -f AppServerAgent.zip AnalyticsAgent.zip)
-  (cd ADCapital-ApplicationProcessor && rm -rf AD-Capital)
-  (cd ADCapital-QueueReader && rm -f AppServerAgent.zip AnalyticsAgent.zip)
-  (cd ADCapital-QueueReader && rm -rf AD-Capital)
-  (cd ADCapital-Load && rm -rf AD-Capital-Load)
-  (cd ADCapital-Java && rm -f jdk-linux-x64.rpm)
+  if [ -z ${PREPARE_ONLY} ]; then 
+    (cd ADCapital-Tomcat && rm -f AppServerAgent.zip AnalyticsAgent.zip env.sh start-analytics.sh)
+    (cd ADCapital-Tomcat && rm -rf AD-Capital)
+    (cd ADCapital-ApplicationProcessor && rm -f AppServerAgent.zip AnalyticsAgent.zip env.sh start-analytics.sh)
+    (cd ADCapital-ApplicationProcessor && rm -rf AD-Capital)
+    (cd ADCapital-QueueReader && rm -f AppServerAgent.zip AnalyticsAgent.zip env.sh start-analytics.sh)
+    (cd ADCapital-QueueReader && rm -rf AD-Capital)
+    (cd ADCapital-Load && rm -rf AD-Capital-Load)
+    (cd ADCapital-Java && rm -f jdk-linux-x64.rpm)
+  fi
 
   # Remove dangling images left-over from build
   if [[ `docker images -q --filter "dangling=true"` ]]
@@ -27,16 +29,69 @@ cleanUp() {
   fi
 }
 trap cleanUp EXIT
+
 promptForAgents() {
   read -e -p "Enter path to App Server Agent: " APP_SERVER_AGENT
   read -e -p "Enter path to Analytics Agent: " ANALYTICS_AGENT
   read -e -p "Enter path to Oracle JDK7: " ORACLE_JDK7
 }
 
+buildContainers() {
+  echo; echo "Building ADCapital-Java..."
+  (cd ADCapital-Java; docker build -t appdynamics/adcapital-java .)
+
+  echo; echo "Building ADCapital-Tomcat..."
+  (cd ADCapital-Tomcat && git clone https://github.com/Appdynamics/AD-Capital.git)
+  (cd ADCapital-Tomcat && docker build -t appdynamics/adcapital-tomcat .)
+
+  echo; echo "Building ADCapital-ApplicationProcessor..."
+  (cd ADCapital-ApplicationProcessor && git clone https://github.com/Appdynamics/AD-Capital.git)
+  (cd ADCapital-ApplicationProcessor && docker build -t appdynamics/adcapital-applicationprocessor .)
+
+  echo; echo "Building ADCapital-QueueReader..."
+  (cd ADCapital-QueueReader && git clone https://github.com/Appdynamics/AD-Capital.git)
+  (cd ADCapital-QueueReader && docker build -t appdynamics/adcapital-queuereader .)
+
+  echo; echo "Building ADCapital-Load..."
+  (cd ADCapital-Load && git clone https://github.com/Appdynamics/AD-Capital-Load.git)
+  (cd ADCapital-Load && docker build -t appdynamics/adcapital-load .)
+}
+
 # Prompt for location of App Server, Machine and Database Agents
 if  [ $# -eq 0 ]
 then
   promptForAgents
+else
+  # Allow user to specify locations of App Server and Analytics Agents
+  while getopts "a:y:j:p:" opt; do
+    case $opt in
+      a)
+        APP_SERVER_AGENT=$OPTARG
+        if [ ! -e ${APP_SERVER_AGENT} ]; then
+          echo "Not found: ${APP_SERVER_AGENT}"; exit
+        fi
+        ;;
+      y)
+        ANALYTICS_AGENT=$OPTARG 
+	if [ ! -e ${ANALYTICS_AGENT} ]; then
+          echo "Not found: ${ANALYTICS_AGENT}"; exit         
+        fi
+        ;;
+      j)
+        ORACLE_JDK7=$OPTARG
+        if [ ! -e ${ORACLE_JDK7} ]; then
+          echo "Not found: ${ORACLE_JDK7}"; exit
+        fi
+        ;; 
+      p)
+        echo "Prepare build environment only - no docker builds"
+        PREPARE_ONLY=true;
+        ;;
+      \?)
+        echo "Invalid option: -$OPTARG"
+        ;;
+    esac
+  done
 fi
 
 if [ -z ${APP_SERVER_AGENT} ]; then
@@ -44,9 +99,8 @@ if [ -z ${APP_SERVER_AGENT} ]; then
 fi
 
 if [ -z ${ANALYTICS_AGENT} ]; then
-    echo "Error: Javascript Agent is required"; exit
+    echo "Error: Analytics Agent is required"; exit
 fi
-
 
 if [ -z ${ORACLE_JDK7} ]
 then
@@ -56,10 +110,6 @@ else
     echo "Using JDK: ${ORACLE_JDK7}"
     cp ${ORACLE_JDK7} ADCapital-Java/jdk-linux-x64.rpm
 fi
-
-echo "Building ADCapital-Java..."
-(cd ADCapital-Java; docker build -t appdynamics/adcapital-java .)
-echo
 
 # If supplied, add standalone analytics agent to build
 if [ -z ${ANALYTICS_AGENT} ]
@@ -72,33 +122,28 @@ else
     cp ${ANALYTICS_AGENT} ADCapital-ApplicationProcessor/AnalyticsAgent.zip
     cp ${ANALYTICS_AGENT} ADCapital-QueueReader/AnalyticsAgent.zip
 
+    cp start-analytics.sh ADCapital-Tomcat
+    cp start-analytics.sh ADCapital-ApplicationProcessor
+    cp start-analytics.sh ADCapital-QueueReader
+
     # Add analytics agent when creating Dockerfile for machine agent
     DOCKERFILE_OPTIONS="analytics"
 fi
 
+echo "Installing App Server Agent"
 echo " ${APP_SERVER_AGENT}"
 cp ${APP_SERVER_AGENT} ADCapital-Tomcat/AppServerAgent.zip
-echo "Copied Agents for ADCapital-Tomcat"
-
-
 cp ${APP_SERVER_AGENT} ADCapital-ApplicationProcessor/AppServerAgent.zip
-echo "Copied Agents for ADCapital-ApplicationProcessor"
-
 cp ${APP_SERVER_AGENT} ADCapital-QueueReader/AppServerAgent.zip
-echo "Copied Agents for ADCapital-QueueReader"
 
-echo; echo "Building ADCapital-Tomcat..."
-(cd ADCapital-Tomcat && git clone https://github.com/Appdynamics/AD-Capital.git)
-(cd ADCapital-Tomcat && docker build -t appdynamics/adcapital-tomcat .)
+echo "Copying environment settings for containers"
+cp env.sh ADCapital-Tomcat
+cp env.sh ADCapital-ApplicationProcessor
+cp env.sh ADCapital-QueueReader
 
-echo; echo "Building ADCapital-ApplicationProcessor..."
-(cd ADCapital-ApplicationProcessor && git clone https://github.com/Appdynamics/AD-Capital.git)
-(cd ADCapital-ApplicationProcessor && docker build -t appdynamics/adcapital-applicationprocessor .)
-
-echo; echo "Building ADCapital-QueueReader..."
-(cd ADCapital-QueueReader && git clone https://github.com/Appdynamics/AD-Capital.git)
-(cd ADCapital-QueueReader && docker build -t appdynamics/adcapital-queuereader .)
-
-echo; echo "Building ADCapital-Load..."
-(cd ADCapital-Load && git clone https://github.com/Appdynamics/AD-Capital-Load.git)
-(cd ADCapital-Load && docker build -t appdynamics/adcapital-load .)
+# Skip build if -p flag (Prepare only) set
+if [ "${PREPARE_ONLY}" = true ] ; then
+    echo "Skipping build phase"
+else
+    buildContainers
+fi
